added error tool tip theme, fixed bug that would display errors on last step prematurely
